+++ packages/llm/src/application/useCases/getLLMConfiguration/index.ts
@@ -1,1 +1,1 @@
-export { GetLLMConfigurationUseCase } from './getLLMConfiguration.usecase';
+export { GetLLMConfigurationUseCase } from './GetLLMConfigurationUseCase';
