+++ packages/llm/src/application/useCases/saveLLMConfiguration/index.ts
@@ -1,1 +1,1 @@
-export { SaveLLMConfigurationUseCase } from './saveLLMConfiguration.usecase';
+export { SaveLLMConfigurationUseCase } from './SaveLLMConfigurationUseCase';
